+++ lib/src/intl_helpers.dart
@@ -19,15 +19,45 @@ typedef MessageIfAbsent(String message_str, List args);
 class UninitializedLocaleData<F> implements MessageLookup {
   final String message;
   final F fallbackData;
-  const UninitializedLocaleData(this.message, this.fallbackData);
+  UninitializedLocaleData(this.message, this.fallbackData);
 
   operator [](String key) =>
       (key == 'en_US') ? fallbackData : _throwException();
 
-  String lookupMessage(String message_str, String locale, String name,
-          List args, String meaning,
-          {MessageIfAbsent ifAbsent}) =>
-      message_str;
+  /// If a message is looked up before any locale initialization, record it,
+  /// and throw an exception with that information once the locale is
+  /// initialized.
+  ///
+  /// Set this during development to find issues with race conditions between
+  /// message caching and locale initialization. If the results of Intl.message
+  /// calls aren't being cached, then this won't help.
+  ///
+  /// There's nothing that actually sets this, so checking this requires
+  /// patching the code here.
+  static final bool throwOnFallback = false;
+
+  /// The messages that were called before the locale was initialized.
+  List<String> _badMessages = [];
+
+  void _reportErrors() {
+    if (throwOnFallback && _badMessages.length > 0) {
+      throw new StateError(
+          "The following messages were called before locale initialization:"
+          " $_uninitializedMessages");
+    }
+  }
+
+  String get _uninitializedMessages =>
+      (_badMessages.toSet().toList()..sort()).join("\n    ");
+
+  String lookupMessage(
+      String message_str, String locale, String name, List args, String meaning,
+      {MessageIfAbsent ifAbsent}) {
+    if (throwOnFallback) {
+      _badMessages.add(name ?? message_str);
+    }
+    return message_str;
+  }
 
   /// Given an initial locale or null, returns the locale that will be used
   /// for messages.
@@ -67,13 +97,15 @@ abstract class LocaleDataReader {
 /// by the implementing package so that we're not dependent on its
 /// implementation.
 MessageLookup messageLookup =
-    const UninitializedLocaleData('initializeMessages(<locale>)', null);
+    new UninitializedLocaleData('initializeMessages(<locale>)', null);
 
 /// Initialize the message lookup mechanism. This is for internal use only.
 /// User applications should import `message_lookup_by_library.dart` and call
 /// `initializeMessages`
 void initializeInternalMessageLookup(Function lookupFunction) {
   if (messageLookup is UninitializedLocaleData) {
+    // This line has to be precisely this way to work around an analyzer crash.
+    (messageLookup as UninitializedLocaleData)._reportErrors();
     messageLookup = lookupFunction();
   }
 }